Login page, login system start Change images->imagens (apparently it's reserved)

+++ models/Image.php
@@ -114,6 +114,10 @@
 
                     $i = 1;
                     $newname = $name;
+
+                    echo $path_to_save;
+
+                    // rename if file exists with same filename
                     while (file_exists($path_to_save)) {
                         if (substr($newname, -2) == "_$i") {
                             $i++;
@@ -188,7 +192,12 @@
             if($this->id && $this->read_single()) {
                 try {
                     $this->url = getUrlDir($this->url);
-                    $deleted = unlink($this->url);
+                    if(!is_file($this->url)) {
+                        $deleted = true;
+                    }
+                    else {
+                        $deleted = unlink($this->url);
+                    }
                 } catch(Exception $e) {
                     return false;
                 }
@@ -202,8 +211,6 @@
                 return true;
             }
 
-            // Print error if something goes wrong
-            printf("Error: %s\n", $stmt->err);
             return false;
         }
     }
